feat: Initial Commit
The majority of the DBus API is now supported. The remaining portion of
the API will be trivial to complete, largely consisting of adding the
missing client conveniences and converting the dynamic values into
static types.

Device and Remote objects are deserialized by hand, due to lack of serde
support in dbus-rs, and all other objects are a hash map of dynamic
values.

An example is included to demonstrate how to use the API.